Enum wpexec :: ModuleType

pub(crate) enum ModuleType {
    Lua,
    LuaConfig,
    Wireplumber,
    Pipewire,
}

The type of module to be loaded

Variants

Lua

A Lua WirePlumber script

LuaConfig

A WirePlumber Lua configuration file

Wireplumber

A native WirePlumber module

Pipewire

A PipeWire module

Implementations

impl ModuleType

pub(crate) fn is_lua (&self) -> bool

pub(crate) fn module_dir (&self) -> Option <&'static str >

Environment variable that affects path resolution for this module type

The wireplumber ComponentLoader generally doesn’t accept absolute paths for modules. This environment variable can be modified in order to help it find the module to load.

pub(crate) fn loader_module (&self) -> Option <&'static str >

A module that provides the necessary loader

While wireplumber has built-in support for loading modules, the Lua scripting engine is itself implemented as a module, which must be loaded first. Afterwards, lua scripts may be loaded as modules (well, components) themselves.

pub(crate) fn loader_type (&self) -> &'static str

The type name expected by a ComponentLoader

This is passed on to Core::load_component .