Struct wpexec::Args

pub(crate) struct Args {
    pub(crate) module_type: ModuleType,
    pub(crate) json_arg: Option<String>,
    pub(crate) plugins: Vec<String>,
    pub(crate) module: Option<String>,
}

Command-line arguments parsed via clap

Fields
module_type: ModuleTypejson_arg: Option<String>

Arguments to pass to the loaded module

Lua scripts only support arrays and dictionary maps.

plugins: Vec<String>

Associated plugins to load, provided by the module

module: Option<String>

Name or full path of the module or script to load

Implementations

impl Args

pub(crate) fn plugins(&self) -> Vec<&str>

The Plugin names to load as part of initialization

For example, the ModuleType::Lua module provides a lua-scripting plugin that’s responsible for loading and running lua scripts.

pub(crate) fn module(&self) -> Option<&str>

The module or script to load

This can be a full path or often just be a name if WirePlumber knows where to find it (the rules for this are convoluted though, so I won’t get into that here)

pub(crate) fn variant(&self) -> Result<Option<LuaVariant<'_>>>

A JSON-like blob of data to pass as an argument to the script or module to be loaded

In practice this must be a dictionary or array because lua scripts can’t work with other types of data as the top-level container, but more specific types may be usable by other modules. See wireplumber::lua for a more detailed explanation.